+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 * **mlab:** battles now expire after 30 minutes instead of 6 hours ([acdc563](https://github.com/IdleLands/IdleLands/commit/acdc563))
 * **movement:** fix a weird case where movement would sometimes not generate valid weights ([255787f](https://github.com/IdleLands/IdleLands/commit/255787f))
 * **party:** allow you to forcibly request your party via frontend, making it visible to semi-login [refreshes] ([4fd5224](https://github.com/IdleLands/IdleLands/commit/4fd5224))
+* **shops:** buying an item will now make a choice ([e10c974](https://github.com/IdleLands/IdleLands/commit/e10c974))
 * **shops:** increment some stats when buying items ([a7faa74](https://github.com/IdleLands/IdleLands/commit/a7faa74))
 * **statistics:** move combat/combatsolo under the combat root ([fe77254](https://github.com/IdleLands/IdleLands/commit/fe77254)), closes [#589](https://github.com/IdleLands/IdleLands/issues/589)