+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 * **collectibles:** always update collectibles when finding a new one ([404e182](https://github.com/IdleLands/IdleLands/commit/404e182)), closes [#610](https://github.com/IdleLands/IdleLands/issues/610)
 * **collectibles:** collectible name will actually be set now ([3716f2b](https://github.com/IdleLands/IdleLands/commit/3716f2b))
 * **collectibles:** show the name when getting a new one ([8210ebf](https://github.com/IdleLands/IdleLands/commit/8210ebf))
+* **combat:** can no longer battle yourself ([4911573](https://github.com/IdleLands/IdleLands/commit/4911573)), closes [#612](https://github.com/IdleLands/IdleLands/issues/612)
 * **equipment:** recalculate stats when equipping items and update the player ([f17f75a](https://github.com/IdleLands/IdleLands/commit/f17f75a))
 * **extchat:** attempt to parse messages better to show discord users ([2369a63](https://github.com/IdleLands/IdleLands/commit/2369a63)), closes [#614](https://github.com/IdleLands/IdleLands/issues/614)
 * **gold:** forsakegold no longer eats all of your gold ([62defbc](https://github.com/IdleLands/IdleLands/commit/62defbc))