+++ CHANGELOG.md
@@ -1,5 +1,5 @@
 <a name="0.0.0"></a>
-# [0.0.0](https://github.com/IdleLands/IdleLands/compare/0.9.0...v0.0.0) (2017-03-18)
+# [0.0.0](https://github.com/IdleLands/IdleLands/compare/0.9.0...v0.0.0) (2017-03-19)
 
 
 ### Bug Fixes